services/mock/requirements.txt
Flask
confluent-kafka
Marshmallow
gunicorn
python-dotenv